BazeIter as part of baze source
Move BazeIter from util to the source/bazefetcher.py file and add
method for creating a BazeIter instance to Bazefetcher.
