+++ camille/util/__init__.py
@@ -10,7 +10,6 @@ Available functions
 * :func:`~camille.util.sn_curve`
 """
 
-from .baze_iterator import BazeIter
 from .resample import resample
 from .sncurves import sn_curve
